Added the sine wave route and switched everything to using floats, to make the maths clearer

+++ app/src/main/java/com/example/chiptune/ui/screens/tetris/TetrisScreen.kt
@@ -104,7 +104,7 @@ fun TetrisScreen(
 
             viewModel.channelList.forEachIndexed { index, ch ->
                 val color = channelColors[index % channelColors.size]
-                val wave = waveformData.channelWaveforms[ch.name] ?: ShortArray(0)
+                val wave = waveformData.channelWaveforms[ch.name] ?: FloatArray(0)
 
                 Card(
                     modifier = Modifier.fillMaxWidth(),
